Improved log messages. Split MksMoveResponse and string converter into its own header so that we don't break Python parsing of MKS_COMMANDS.hpp
